# FTP Connection Management Optimization using Singleton Pattern