--- a/pysus/ftp/__init__.py
+++ b/pysus/ftp/__init__.py
@@ -22,10 +22,9 @@ from typing import (
 import humanize
 from aioftp import Client
 from loguru import logger
+from pysus.data.local import Data
 from tqdm import tqdm
 from typing_extensions import Self
-
-from pysus.data.local import Data
 
 # Type aliases
 PathLike = Union[str, pathlib.Path]
@@ -403,7 +402,9 @@ def load_directory_content(path: str) -> FileContent:
         def line_parser(line: str):
             if "<DIR>" in line:
                 date, time, _, name = line.strip().split(maxsplit=3)
-                modify = datetime.strptime(f"{date} {time}", "%m-%d-%y %I:%M%p")
+                modify = datetime.strptime(
+                    f"{date} {time}", "%m-%d-%y %I:%M%p"
+                )
                 info = {"size": 0, "type": "dir", "modify": modify}
                 xpath = f"{path}/{name}"
                 content[name] = Directory(xpath)
@@ -481,7 +482,9 @@ class Database:
         inside content, `load()` the directory and call `content` again.
         """
         if not self.__content__:
-            logger.info("content is not loaded, use `load()` to load default paths")
+            logger.info(
+                "content is not loaded, use `load()` to load default paths"
+            )
             return []
         return sorted(list(self.__content__.values()), key=str)
 
@@ -546,7 +549,9 @@ class Database:
         """
         ...
 
-    def download(self, files: List[File], local_dir: str = CACHEPATH) -> List[str]:
+    def download(
+        self, files: List[File], local_dir: str = CACHEPATH
+    ) -> List[str]:
         """
         Downloads a list of Files.
         """
@@ -561,7 +566,9 @@ class Database:
             return dfiles[0]
         return dfiles
 
-    async def async_download(self, files: List[File], local_dir: str = CACHEPATH):
+    async def async_download(
+        self, files: List[File], local_dir: str = CACHEPATH
+    ):
         """
         Asynchronously downloads a list of files
         """