--- a/pysus/ftp/__init__.py
+++ b/pysus/ftp/__init__.py
@@ -22,9 +22,10 @@ from typing import (
 import humanize
 from aioftp import Client
 from loguru import logger
-from pysus.data.local import Data
 from tqdm import tqdm
 from typing_extensions import Self
+
+from pysus.data.local import Data
 
 # Type aliases
 PathLike = Union[str, pathlib.Path]
@@ -124,6 +125,8 @@ class File:
             Static method to parse a line from the FTP LIST command and
             extract file information.
     """
+
+    """FTP File representation with improved type safety"""
 
     def __init__(self, path: str, name: str, info: FileInfo) -> None:
         self.name, self.extension = os.path.splitext(name)
@@ -372,12 +375,6 @@ class Directory:
         self.loaded = False
         return self.load()
 
-    def _init_root_child(self, name: str) -> None:
-        self.parent = DIRECTORY_CACHE["/"]
-        self.name = name
-        self.loaded = False
-        self.__content__ = {}
-
     def __repr__(self) -> str:
         return self.path
 
@@ -402,9 +399,7 @@ def load_directory_content(path: str) -> FileContent:
         def line_parser(line: str):
             if "<DIR>" in line:
                 date, time, _, name = line.strip().split(maxsplit=3)
-                modify = datetime.strptime(
-                    f"{date} {time}", "%m-%d-%y %I:%M%p"
-                )
+                modify = datetime.strptime(f"{date} {time}", "%m-%d-%y %I:%M%p")
                 info = {"size": 0, "type": "dir", "modify": modify}
                 xpath = f"{path}/{name}"
                 content[name] = Directory(xpath)
@@ -482,9 +477,7 @@ class Database:
         inside content, `load()` the directory and call `content` again.
         """
         if not self.__content__:
-            logger.info(
-                "content is not loaded, use `load()` to load default paths"
-            )
+            logger.info("content is not loaded, use `load()` to load default paths")
             return []
         return sorted(list(self.__content__.values()), key=str)
 
@@ -549,9 +542,7 @@ class Database:
         """
         ...
 
-    def download(
-        self, files: List[File], local_dir: str = CACHEPATH
-    ) -> List[str]:
+    def download(self, files: List[File], local_dir: str = CACHEPATH) -> List[str]:
         """
         Downloads a list of Files.
         """
@@ -566,9 +557,7 @@ class Database:
             return dfiles[0]
         return dfiles
 
-    async def async_download(
-        self, files: List[File], local_dir: str = CACHEPATH
-    ):
+    async def async_download(self, files: List[File], local_dir: str = CACHEPATH):
         """
         Asynchronously downloads a list of files
         """