--- a/pysus/ftp/__init__.py
+++ b/pysus/ftp/__init__.py
@@ -375,6 +375,9 @@ class Directory:
         self.loaded = False
         return self.load()
 
+    def __str__(self) -> str:
+        return self.path
+
     def __repr__(self) -> str:
         return self.path