## Overview
This PR implements an FTP connection singleton pattern to significantly improve module import performance. The change was motivated by discussion #211 where multiple redundant FTP connections were identified as a performance bottleneck.

## Changes
- Implemented `FTPSingleton` class to manage a single FTP connection
- Modified `Directory` class to use the singleton pattern
- Reduced redundant FTP connections across database modules
- Maintained all existing functionality while improving performance

## Performance Impact
The implementation shows dramatic performance improvements:

![image](https://github.com/user-attachments/assets/168ff407-9f93-4b27-822d-78a35fea6d0f)

**Before (main branch):**
- Total import time: 21.0434s
- Individual module imports taking 1.1-4.4s each
- Multiple redundant FTP connections being created

**After (with FTP singleton):**
- Total import time: 0.3499s (98.3% reduction)
- First import: 0.3479s (creates singleton)
- Subsequent imports: ~0.0003s (reuses connection)

### Performance Test Code
```python
import importlib
from time import perf_counter
from typing import Dict
from loguru import logger

def measure_import_time(module_name: str) -> float:
    """Measure import time for a single module"""
    start = perf_counter()
    importlib.import_module(module_name)
    end = perf_counter()
    return end - start

def main():
    # List of modules to test that use Directory class
    modules = [
        "pysus.ftp.databases.sim",
        "pysus.ftp.databases.sia",
        "pysus.ftp.databases.sih",
        "pysus.ftp.databases.sinan",
        "pysus.ftp.databases.sinasc",
        "pysus.ftp.databases.cnes",
        "pysus.ftp.databases.pni",
        "pysus.ftp.databases.ibge_datasus",
    ]

    results: Dict[str, float] = {}
    total_time = 0.0

    logger.info("Starting import measurements...")

    for module in modules:
        try:
            time_taken = measure_import_time(module)
            results[module] = time_taken
            total_time += time_taken
            logger.info(f"{module}: {time_taken:.4f}s")
        except Exception as e:
            logger.error(f"Error importing {module}: {e}")

    logger.info("=" * 50)
    logger.info(f"Total import time: {total_time:.4f}s")
    logger.info("=" * 50)

if __name__ == "__main__":
    main()
```

I believe this change significantly improves the library's performance while maintaining its ease of use and reliability.

Note: I've added some type-hints that was lacking, a lot of warnings was appearing (some of them was solved putting a type ignore alias.)