@felipeadeildo 2024-11-25
Oh, i agree!
Ive removed the content append when rewriting the whole file.
I'm implementing the content loading and removing the unused line parser with operational system conditional (Ive noticed that we don't need it)