@luabida 2024-11-26
Hello @felipeadeildo, could you please rebase your PR based on the latest release? I've fixed the CI and added some endpoint tests that will ensure the ftp functionality. Also, please note that you should be using semantic release commit message format in at least one commit and/or PR title to include your changes in the next release