@felipeadeildo 2024-11-29
Hey!

I'm waiting the ftp server turns on again.

When the service comes back, I will execute the tests (`make test-pysus`) that i've been using to confirm if is everything fine.

So, i will mark this PR as "still in progress" for a while.