@felipeadeildo 2024-12-01
Hi! I've executed the tests, although there are some warnings about deprecated functions from `tqdm` library, all the ftp tests has been successfully passed.

```
pysus/tests/test_ftp.py::TestDirectoryAndFile::test_directory_cache PASSED                                                                                                                       [ 41%]
pysus/tests/test_ftp.py::TestDirectoryAndFile::test_root_directory PASSED                                                                                                                        [ 44%]
pysus/tests/test_ftp.py::TestDirectoryAndFile::test_root_load PASSED                                                                                                                             [ 47%]
pysus/tests/test_ftp.py::TestDirectoryAndFile::test_root_reload PASSED                                                                                                                           [ 50%]
pysus/tests/test_ftp.py::TestDirectoryAndFile::test_sinan_file PASSED                                                                                                                            [ 52%]
pysus/tests/test_ftp.py::TestDatabases::test_ciha PASSED                                                                                                                                         [ 55%]
pysus/tests/test_ftp.py::TestDatabases::test_cnes PASSED                                                                                                                                         [ 58%]
pysus/tests/test_ftp.py::TestDatabases::test_ibge_datasus PASSED                                                                                                                                 [ 61%]
pysus/tests/test_ftp.py::TestDatabases::test_pni PASSED                                                                                                                                          [ 63%]
pysus/tests/test_ftp.py::TestDatabases::test_sia PASSED                                                                                                                                          [ 66%]
pysus/tests/test_ftp.py::TestDatabases::test_sih PASSED                                                                                                                                          [ 69%]
pysus/tests/test_ftp.py::TestDatabases::test_sim PASSED                                                                                                                                          [ 72%]
pysus/tests/test_ftp.py::TestDatabases::test_sinan PASSED                                                                                                                                        [ 75%]
pysus/tests/test_ftp.py::TestDatabases::test_sinasc PASSED
```