@luabida 2024-12-02
@felipeadeildo you may run the command `pre-commit run --all-files` and fix the errors in order to pass on CI.